# docs/swagger.yaml
definitions:
...
  handler.UploadImageResponse:
    properties:
      path:
        type: string
      url:
        type: string
    type: object
...
  request.CreateBlogRequest:
    properties:
      category:
        maxLength: 100
        type: string
      content:
        type: string
      excerpt:
        maxLength: 500
        type: string
      image:
        maxLength: 255
        type: string
      is_published:
        type: boolean
      slug:
        maxLength: 255
        type: string
      tags:
        maxLength: 255
        type: string
      title:
        maxLength: 255
        type: string
    required:
    - content
    - slug
    - title
    type: object
...
  request.UpdateBlogRequest:
    properties:
      category:
        maxLength: 100
        type: string
      content:
        type: string
      excerpt:
        maxLength: 500
        type: string
      image:
        maxLength: 255
        type: string
      is_published:
        type: boolean
      slug:
        maxLength: 255
        type: string
      tags:
        maxLength: 255
        type: string
      title:
        maxLength: 255
        type: string
    required:
    - content
    - slug
    - title
    type: object
...
  response.BlogResponse:
    properties:
      author:
        type: string
      category:
        type: string
      content:
        type: string
      created_at:
        type: string
      excerpt:
        type: string
      id:
        type: integer
      image:
        type: string
      is_published:
        type: boolean
      published_at:
        type: string
      slug:
        type: string
      tags:
        type: string
      title:
        type: string
      updated_at:
        type: string
      view_count:
        type: integer
    type: object
...
info:
  contact: {}
paths:
  /api/v1/admin/blogs:
    get:
      description: Retrieve a list of all blogs with pagination and optional search
      parameters:
      - default: 1
        description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - default: 10
        description: 'Page size (default: 10, max: 100)'
        in: query
        name: page_size
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of blogs with pagination info
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Get all blogs (admin only)
      tags:
      - Blogs
    post:
      consumes:
      - application/json
      description: Create a new blog post (admin only)
      parameters:
      - description: Blog creation request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.CreateBlogRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Blog created successfully
          schema:
            $ref: '#/definitions/response.BlogResponse'
        "400":
          description: Invalid request body
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Create a new blog
      tags:
      - Blogs
  /api/v1/admin/blogs/{id}:
    delete:
      description: Delete a blog post (admin only)
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Blog deleted successfully
        "400":
          description: Invalid blog ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Blog not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Delete a blog
      tags:
      - Blogs
    get:
      description: Retrieve a specific blog by its ID
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Blog details
          schema:
            $ref: '#/definitions/response.BlogResponse'
        "400":
          description: Invalid blog ID
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Blog not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Get blog by ID (admin only)
      tags:
      - Blogs
    put:
      consumes:
      - application/json
      description: Update an existing blog post (admin only)
      parameters:
      - description: Blog ID
        in: path
        name: id
        required: true
        type: integer
      - description: Blog update request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/request.UpdateBlogRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Blog updated successfully
          schema:
            $ref: '#/definitions/response.BlogResponse'
        "400":
          description: Invalid request or blog ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Blog not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      security:
      - Bearer: []
      summary: Update a blog
      tags:
      - Blogs
  /api/v1/admin/blogs/upload/image:
    post:
      consumes:
      - multipart/form-data
      description: 'Upload an image file for a blog post (admin only). Maximum file
        size: 5MB. Supported formats: JPEG, PNG, GIF, WebP'
      parameters:
      - description: Image file to upload
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Image uploaded successfully
          schema:
            $ref: '#/definitions/handler.UploadImageResponse'
        "400":
          description: Invalid file or file too large
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to upload file
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: Upload blog image
      tags:
      - Blogs
...
  /api/v1/blogs:
    get:
      description: Retrieve a list of all published blogs with pagination and optional
        search
      parameters:
      - default: 1
        description: 'Page number (default: 1)'
        in: query
        name: page
        type: integer
      - default: 10
        description: 'Page size (default: 10)'
        in: query
        name: page_size
        type: integer
      - description: Search term
        in: query
        name: search
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of published blogs with pagination info
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get all published blogs
      tags:
      - Blogs
  /api/v1/blogs/{slug}:
    get:
      description: Retrieve a specific blog by its slug (public endpoint)
      parameters:
      - description: Blog slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Blog details with author information
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid slug
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Blog not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Get blog by slug
      tags:
      - Blogs
...
  /images/{filename}:
    get:
      description: Retrieve a blog image by filename (public endpoint)
      parameters:
      - description: Image filename
        in: path
        name: filename
        required: true
        type: string
      produces:
      - image/jpeg
      - image/png
      - image/gif
      - image/webp
      responses:
        "200":
          description: Image file
        "400":
          description: Invalid filename
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Image not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Failed to retrieve image
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get blog image
      tags:
      - Blogs
swagger: "2.0"